requirements.txt
blinker==1.7.0
click==8.1.7
flask==3.0.0
Flask-MQTT==1.1.1
importlib-metadata==7.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
paho-mqtt==1.6.1
werkzeug==3.0.1
zipp==3.17.0
